+++ lib/web/pages/tracing/page.ex
@@ -72,18 +72,18 @@ defmodule Observer.Web.Tracing.Page do
       |> assign(attention_msg: attention_msg)
 
     ~H"""
-    <div class="min-h-screen bg-white">
+    <div class="min-h-screen bg-white dark:bg-gray-800">
       <Attention.content
         id="tracing"
         title="Attention"
-        class="border-red-400 text-red-500"
+        class="border-red-400 dark:border-red-700 text-red-500 dark:text-red-200"
         message={@attention_msg}
       >
         <:inner_form>
           <.form
             for={@form}
             id="tracing-update-form"
-            class="flex ml-2 mr-2 text-xs text-center whitespace-nowrap gap-5"
+            class="flex ml-2 mr-2 text-xs text-center text-zinc-800 dark:text-white  whitespace-nowrap gap-5"
             phx-change="form-update"
           >
             <Core.input
@@ -110,7 +110,7 @@ defmodule Observer.Web.Tracing.Page do
             :if={@trace_idle? and @trace_owner?}
             id="tracing-multi-select-run"
             phx-click="tracing-apps-run"
-            class="phx-submit-loading:opacity-75 rounded-r-xl bg-green-500 transform active:scale-75 transition-transform hover:bg-green-600 py-10 w-64 text-sm font-semibold  text-white active:text-white/80"
+            class="phx-submit-loading:opacity-75 rounded-r-xl bg-green-500 dark:bg-green-700 transform active:scale-75 transition-transform hover:bg-green-800 dark:hover:bg-green-800 py-10 w-64 text-sm font-semibold text-white active:text-white/80"
           >
             RUN
           </button>
@@ -118,14 +118,14 @@ defmodule Observer.Web.Tracing.Page do
             :if={@trace_idle? == false and @trace_owner?}
             id="tracing-multi-select-stop"
             phx-click="tracing-apps-stop"
-            class="phx-submit-loading:opacity-75 rounded-r-xl bg-red-500 transform active:scale-75 transition-transform hover:bg-red-600 py-10 w-64 text-sm font-semibold text-white active:text-white/80 animate-pulse"
+            class="phx-submit-loading:opacity-75 rounded-r-xl bg-red-500 dark:bg-red-700 transform active:scale-75 transition-transform hover:bg-red-600 dark:hover:bg-red-800 py-10 w-64 text-sm font-semibold text-white active:text-white/80 animate-pulse"
           >
             STOP
           </button>
 
           <button
             :if={not @trace_owner?}
-            class="phx-submit-loading:opacity-75 rounded-r-xl bg-red-500 transform active:scale-75 transition-transform hover:bg-red-600 py-10 w-64 text-sm font-semibold text-white active:text-white/80 animate-pulse"
+            class="phx-submit-loading:opacity-75 rounded-r-xl bg-red-500 dark:bg-red-700 transform active:scale-75 transition-transform hover:bg-red-600 dark:hover:bg-red-800 py-10 w-64 text-sm font-semibold text-white active:text-white/80 animate-pulse"
           >
             IN USE
           </button>
@@ -152,7 +152,7 @@ defmodule Observer.Web.Tracing.Page do
         />
       </div>
       <div class="p-2">
-        <div class="bg-white w-full shadow-lg rounded">
+        <div class="bg-white dark:bg-gray-800 w-full shadow-lg rounded">
           <Core.table_tracing id="live-logs" rows={@streams.tracing_messages}>
             <:col :let={{_id, tracing_message}} label="SERVICE">
               <span>{tracing_message.service}</span>